final bits of content changes for v2.1
- start refactoring gameState
- more quests
- add some unit tests
