+++ frontend/package.json
@@ -12,8 +12,11 @@
   },
   "devDependencies": {
     "@astrojs/svelte": "^2.1.0",
+    "@babel/core": "^7.22.9",
+    "@babel/preset-env": "^7.22.9",
     "ajv": "^8.12.0",
     "astro": "^2.3.2",
+    "babel-jest": "^29.6.1",
     "jest": "^29.4.3",
     "svelte": "^3.50.1",
     "svelte-switch-case": "^2.0.0"
@@ -44,6 +47,6 @@
     "postcss": "^8.4.23",
     "pretty-date": "^0.2.0",
     "remarkable": "^2.0.1",
-    "web3": "^0.20.7"
+    "web3": "^4.0.3"
   }
 }
