+++ frontend/src/utils/gameState.js
@@ -1,53 +1,5 @@
-import { writable } from 'svelte/store';
-import processes from '../pages/processes/processes.json';
-import { durationInSeconds } from '../utils.js';
-
-// ---------------------
-// INTERNAL: GAME STATE
-// ---------------------
-
-const gameStateKey = 'gameState';
-
-const initializeGameState = () => {
-  return {
-    quests: {},
-    inventory: {},
-    processes: {},
-  };
-};
-
-export const loadGameState = () => {
-  const storedGameState = localStorage.getItem(gameStateKey);
-  if (storedGameState) {
-    return JSON.parse(storedGameState);
-  }
-  return initializeGameState();
-};
-
-export const saveGameState = (gameState) => {
-  localStorage.setItem(gameStateKey, JSON.stringify(gameState));
-  state.set(gameState); // Update the state store directly
-};
-
-let gameState = loadGameState();
-
-// Create the state store and set the initial value
-export const state = writable(gameState);
-
-export const exportGameStateString = () => {
-  return btoa(JSON.stringify(gameState));
-};
-
-export const importGameStateString = (gameStateString) => {
-  // Decode from Base64 and parse the JSON string
-  const importedGameState = JSON.parse(atob(gameStateString));
-
-  // Overwrite the game state with the imported game state
-  gameState = importedGameState;
-
-  // Save the updated game state to local storage and update the state store
-  saveGameState(gameState);
-};
+import { loadGameState, saveGameState,  } from "./gameState/common.js";
+import { burnItems, addItems, getItemCounts, getItemCount, getCurrentdUSD, buyItems } from "./gameState/inventory.js";
 
 // ---------------------
 // QUESTS
@@ -114,89 +66,7 @@ export const grantItems = (questId, stepId, optionIndex, itemList) => {
   }
   addItems(itemList);
   setItemsGranted(questId, stepId, optionIndex);
-}
-
-// ---------------------
-// INVENTORY
-// ---------------------
-
-export const addItems = (items) => {
-  items.forEach(({ id, count }) => {
-    gameState.inventory[id] = (gameState.inventory[id] || 0) + count;
-  });
-  saveGameState(gameState);
 };
-
-export const burnItems = (items) => {
-  items.forEach(({ id, count }) => {
-    if (gameState.inventory[id]) {
-      gameState.inventory[id] -= count;
-      if (gameState.inventory[id] < 0) {
-        gameState.inventory[id] = 0;
-      }
-    }
-  });
-  saveGameState(gameState);
-};
-
-export const getItemCounts = (itemList) => {
-  const counts = {};
-  itemList.forEach((item) => {
-    counts[item.id] = gameState.inventory[item.id] || 0;
-  });
-  return counts;
-};
-
-export const getItemCount = (itemId) => {
-  return getItemCounts([{ id: itemId }])[itemId];
-};
-
-export const getCurrentdUSD = () => {
-  return getItemCount("24");
-};
-
-export const buyItems = (items) => {
-  items.forEach(item => {
-    const { price, quantity } = item;
-    const currencyId = "24"; // Assuming the currency ID is always "24"
-
-    const totalPrice = parseFloat(price) * parseFloat(quantity); // Convert price and quantity to numbers
-
-    if (gameState.inventory[currencyId] && gameState.inventory[currencyId] >= totalPrice) {
-      gameState.inventory[currencyId] -= totalPrice;
-      gameState.inventory[item.id] = (gameState.inventory[item.id] || 0) + parseFloat(quantity);
-    }
-  });
-
-  saveGameState(gameState);
-};
-
-export const sellItems = (items) => {
-  items.forEach(item => {
-    const { price, quantity } = item; // get the quantity from the item
-    const currencyId = "24"; // Assuming the currency ID is always "24"
-
-    if (gameState.inventory[item.id] && gameState.inventory[item.id] >= quantity) {
-      gameState.inventory[item.id] -= quantity;
-      gameState.inventory[currencyId] = (gameState.inventory[currencyId] || 0) + price * quantity;
-    }
-  });
-
-  saveGameState(gameState);
-};
-
-
-
-export const hasItems = (itemList) => {
-  for (let i = 0; i < itemList.length; i++) {
-    const { id, count } = itemList[i];
-    if (!gameState.inventory[id] || gameState.inventory[id] < count) {
-      return false;
-    }
-  }
-  return true;
-};
-
 
 // ---------------------
 // PROCESSES
